# docs/api/erp-v1-openapi.yaml
openapi: 3.1.0
info:
  title: ERP Operational Core API
  version: 1.0.0
  description: "Phase 1 production contract. Lifecycle and scope live in operational-core-v1-spec.md; shared rules live in api-conventions.md."
servers:
  - url: https://erp.example.invalid
    description: "Same-site reverse proxy that forwards /api to the backend."
paths:
...
  /api/v1/production-orders:
    get:
      operationId: listProductionOrders
      summary: "List Production Orders"
      description: "Filters: status, buyerOrderId, productionNo, productKind, groupNo. Sort allowlist is productionNo, status, createdAt and id."
      security: [{ bearerAuth: [] }]
      x-required-permission: "PRODUCTION:VIEW"
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - { name: status, in: query, schema: { type: string, enum: [OPEN, FINISHED, CANCELLED] } }
        - { name: buyerOrderId, in: query, schema: { type: string, format: uuid } }
        - { name: productionNo, in: query, schema: { type: string } }
        - { name: productKind, in: query, schema: { type: string, enum: [PRINT, WOVEN] } }
        - { name: groupNo, in: query, schema: { type: string } }
      responses:
        '200':
          description: "Paged Production Orders."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductionOrderPageResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/production-orders/{id}:
    get:
      operationId: getProductionOrder
      summary: "Read one Production Order with configuration and group"
      security: [{ bearerAuth: [] }]
      x-required-permission: "PRODUCTION:VIEW"
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: "Production Order detail."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductionOrderResponse' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/production-orders/{id}/configuration:
    put:
      operationId: configureProductionOrder
      summary: "Replace the matching PRINT or WOVEN configuration while OPEN"
      security: [{ bearerAuth: [] }]
      x-required-permission: "PRODUCTION:CONFIGURE"
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductionConfigurationRequest' } } }
      responses:
        '200':
          description: "Canonical configuration saved and event appended."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductionOrderResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/CommandConflict' }
        '410': { $ref: '#/components/responses/IdempotencyResultExpired' }
        '422': { $ref: '#/components/responses/UnprocessableCommand' }
  /api/v1/production-orders/{id}/finish:
    post:
      operationId: finishProductionOrder
      summary: "Finish Production once and atomically create FG stock"
      description: "The command accepts no produced quantity; produced quantity equals planned quantity."
      security: [{ bearerAuth: [] }]
      x-required-permission: "PRODUCTION:FINISH"
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductionCommandRequest' } } }
      responses:
        '200':
          description: "Finished Production and reconciled Stock Position."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductionFinishResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/CommandConflict' }
        '410': { $ref: '#/components/responses/IdempotencyResultExpired' }
        '422': { $ref: '#/components/responses/UnprocessableCommand' }
  /api/v1/production-orders/{id}/events:
    get:
      operationId: listProductionEvents
      summary: "Read append-only Production history"
      security: [{ bearerAuth: [] }]
      x-required-permission: "PRODUCTION:VIEW"
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: "Sort allowlist is occurredAt and id."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductionEventPageResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/production-orders/{id}/document.pdf:
    get:
      operationId: downloadProductionDocument
      summary: "Stream the Production PDF for OPEN or FINISHED status"
      description: "Generated synchronously from canonical data, explicitly labels status, and is never persisted."
      security: [{ bearerAuth: [] }]
      x-required-permission: "PRODUCTION:VIEW"
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: "Production PDF."
          headers:
            Content-Disposition: { $ref: '#/components/headers/ContentDisposition' }
            Cache-Control: { $ref: '#/components/headers/NoStore' }
          content: { application/pdf: { schema: { type: string, format: binary } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/ReportBusy' }
  /api/v1/production-groups:
    post:
      operationId: createProductionGroup
      summary: "Group 2 to 50 OPEN Production Orders from one Buyer Order"
      description: "Each unique member carries its own expectedVersion. groupNo is server-generated as PG-<SYS-PO>-NNN."
      security: [{ bearerAuth: [] }]
      x-required-permission: "PRODUCTION:GROUP"
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductionGroupCreateRequest' } } }
      responses:
        '201':
          description: "Group created with canonical members."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductionGroupResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/CommandConflict' }
        '410': { $ref: '#/components/responses/IdempotencyResultExpired' }
        '422': { $ref: '#/components/responses/UnprocessableCommand' }
  /api/v1/production-groups/{id}/ungroup:
    post:
      operationId: ungroupProductionGroup
      summary: "Ungroup an OPEN Production Group"
      security: [{ bearerAuth: [] }]
      x-required-permission: "PRODUCTION:GROUP"
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProductionCommandRequest' } } }
      responses:
        '200':
          description: "Group removed and member events appended."
          headers: { Cache-Control: { $ref: '#/components/headers/NoStore' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ProductionGroupResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/CommandConflict' }
        '410': { $ref: '#/components/responses/IdempotencyResultExpired' }
        '422': { $ref: '#/components/responses/UnprocessableCommand' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "RS256 access JWT signed by an RSA-3072 key with a kid header and a 15 minute lifetime."
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: "Required for unsafe cookie-backed Auth calls. Reacquire it after login and logout."
    refreshCookie:
      type: apiKey
      in: cookie
      name: ERP_REFRESH
      description: "Rotating opaque refresh token. Secure HttpOnly SameSite=Strict and scoped to /api/v1/auth. Only its hash is stored server-side."
  headers:
    NoStore:
      description: "Always no-store for protected transaction and report responses."
      schema: { type: string, enum: [no-store] }
    ContentDisposition:
      description: "Attachment filename sanitized by the server; no client-controlled path or markup."
      schema: { type: string }
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: "Opaque client retry key, scoped by actor, method, and normalized path. The request hash includes body and If-Match. Retained for 24 hours."
      schema: { type: string, minLength: 1, maxLength: 120 }
    Page:
      name: page
      in: query
      description: "Zero-based page index."
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      description: "Page size."
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Sort:
      name: sort
      in: query
      description: "Repeatable allowlisted sort term written as field,direction. Repeat the parameter to add a secondary term such as a stable id tiebreaker."
      style: form
      explode: true
      schema:
        type: array
        items: { type: string, example: "code,asc" }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: "Quoted optimistic version of the addressed resource."
      schema: { type: string, pattern: '^"[0-9]+"$' }
  responses:
    BadRequest:
      description: "INVALID_REQUEST or VALIDATION_FAILED, including unknown request members."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Unauthenticated:
      description: "UNAUTHENTICATED, CSRF_INVALID or REFRESH_REAUTH_REQUIRED. The SPA may attempt one refresh for these codes."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Forbidden:
      description: "FORBIDDEN when effective permissions deny the operation, or PASSWORD_CHANGE_REQUIRED when a restricted challenge calls anything other than change-password or logout. Neither code may trigger a refresh attempt."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: "NOT_FOUND for the addressed resource or ownership path."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    CommandConflict:
      description: "VERSION_CONFLICT, IDEMPOTENCY_IN_PROGRESS, invalid transition, downstream activity, stock conflict, or another stable command conflict."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    IdempotencyResultExpired:
      description: "A quarantined legacy same-hash result cannot be replayed until its record is purged."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    UnprocessableCommand:
      description: "IDEMPOTENCY_KEY_REUSED or another semantically invalid command."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
    ReportBusy:
      description: "Both report generation slots are occupied on this single backend instance."
      headers:
        Retry-After: { schema: { type: integer, enum: [5] } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
    DecimalQuantity:
      type: string
      pattern: '^\d{1,14}(\.\d{1,4})?$'
      description: "Non-negative decimal quantity carried as a JSON string with up to four decimal places."
...
    ProblemDetail:
      type: object
      additionalProperties: true
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: "Stable machine-readable error code."
          enum:
            - INVALID_REQUEST
            - VALIDATION_FAILED
            - UNAUTHENTICATED
            - CSRF_INVALID
            - REFRESH_REAUTH_REQUIRED
            - PASSWORD_CHANGE_REQUIRED
            - FORBIDDEN
            - NOT_FOUND
            - VERSION_CONFLICT
            - DUPLICATE_BUSINESS_KEY
            - MASTER_IN_USE
            - RECOVERY_ADMIN_REQUIRED
            - EXCHANGE_RATE_MISSING
            - RATE_LIMITED
            - INTERNAL_ERROR
            - IF_MATCH_REQUIRED
            - INVALID_IF_MATCH
            - IDEMPOTENCY_KEY_REQUIRED
            - INVALID_IDEMPOTENCY_KEY
            - IDEMPOTENCY_KEY_REUSED
            - IDEMPOTENCY_IN_PROGRESS
            - IDEMPOTENCY_RESULT_EXPIRED
            - INVALID_STATE_TRANSITION
            - DOWNSTREAM_ACTIVITY_EXISTS
            - GROUP_MEMBERSHIP_INVALID
            - PRODUCTION_ALREADY_FINISHED
            - INSUFFICIENT_STOCK
            - DELIVERY_ALREADY_POSTED
            - DELIVERY_ALREADY_REVERSED
            - DELIVERY_HAS_RETURNS
            - RETURN_LIMIT_EXCEEDED
            - DISPOSAL_EXCEEDS_STOCK
            - REPORT_LIMIT_EXCEEDED
            - REPORT_BUSY
        traceId: { type: string }
        fieldErrors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
...
    PageMetadata:
      type: object
      additionalProperties: false
      required: [number, size, totalElements, totalPages]
      properties:
        number: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    AppliedFilters:
      type: object
      additionalProperties: true
      description: "Echo of the accepted allowlisted filters only."
    AppliedSort:
      type: array
      items: { type: string }
      description: "Echo of the normalized allowlisted sort terms."
...
    DecimalPercent:
      type: string
      pattern: '^(100(\.0{1,4})?|([0-9]|[1-9][0-9])(\.\d{1,4})?)$'
      description: "Percentage from zero through one hundred, with up to four decimal places."
...
    ProductionGroupMemberRequest:
      type: object
      additionalProperties: false
      required: [productionOrderId, expectedVersion]
      properties:
        productionOrderId: { type: string, format: uuid }
        expectedVersion: { type: integer, minimum: 0 }
    ProductionGroupCreateRequest:
      type: object
      additionalProperties: false
      required: [members]
      properties:
        members:
          type: array
          minItems: 2
          maxItems: 50
          uniqueItems: true
          items: { $ref: '#/components/schemas/ProductionGroupMemberRequest' }
        reason: { type: string, maxLength: 200 }
    ProductionProcessRequest:
      type: object
      additionalProperties: false
      required: [processId, sequenceNo]
      properties:
        processId: { type: string, format: uuid }
        sequenceNo: { type: integer, minimum: 1 }
        speed: { $ref: '#/components/schemas/DecimalQuantity' }
    ProductionYarnLineRequest:
      type: object
      additionalProperties: false
      properties:
        yarn: { type: string, maxLength: 120 }
        yarnCode: { type: string, maxLength: 120 }
        denier: { type: string, maxLength: 120 }
    ProductionConfigurationRequest:
      type: object
      additionalProperties: false
      required: [productKind, processes, yarnLines]
      properties:
        productKind: { type: string, enum: [PRINT, WOVEN] }
        materialSource: { type: string, enum: [STOCK, PURCHASE] }
        orderKind: { type: string, enum: [LAYOUT, SAMPLE, NEW_ORDER, REPEAT_ORDER, SECOND_PRODUCTION] }
        bim: { type: string, enum: [WHITE, BLACK, 100D, 75D, 50D] }
        tenDia: { type: string, maxLength: 120 }
        matDo: { type: string, maxLength: 120 }
        pick: { type: string, maxLength: 120 }
        xNgangMm: { $ref: '#/components/schemas/DecimalQuantity' }
        yDocMm: { $ref: '#/components/schemas/DecimalQuantity' }
        haiDaMm: { $ref: '#/components/schemas/DecimalQuantity' }
        logoXMm: { $ref: '#/components/schemas/DecimalQuantity' }
        logoYMm: { $ref: '#/components/schemas/DecimalQuantity' }
        hoPercent: { $ref: '#/components/schemas/DecimalPercent' }
        weaveTypes: { type: array, uniqueItems: true, items: { type: string, enum: [THUONG, HAI_DA, SATIN, SATIN_HAI_DA] } }
        processes:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/ProductionProcessRequest' }
        yarnLines:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/ProductionYarnLineRequest' }
        remark: { type: string, maxLength: 200 }
    ProductionCommandRequest:
      type: object
      additionalProperties: false
      properties:
        reason: { type: string, maxLength: 200 }
    ProductionEventResponse:
      type: object
      additionalProperties: false
      required: [id, eventType, actorUserId, occurredAt]
      properties:
        id: { type: string, format: uuid }
        eventType: { type: string, enum: [CREATED, GROUPED, UNGROUPED, CONFIGURED, FINISHED, CANCELLED] }
        actorUserId: { type: string, format: uuid }
        reason: { type: string }
        occurredAt: { type: string, format: date-time }
    ProductionOrderResponse:
      type: object
      additionalProperties: false
      required: [id, version, productionNo, productNo, qrValue, buyerOrderId, buyerOrderItemId, productKind, plannedQty, status, configuration, events, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 0 }
        productionNo: { type: string }
        productNo: { type: string }
        qrValue: { type: string }
        groupId: { type: string, format: uuid }
        groupNo: { type: string }
        buyerOrderId: { type: string, format: uuid }
        buyerOrderItemId: { type: string, format: uuid }
        productKind: { type: string, enum: [PRINT, WOVEN] }
        plannedQty: { $ref: '#/components/schemas/DecimalQuantity' }
        producedQty: { $ref: '#/components/schemas/DecimalQuantity' }
        status: { type: string, enum: [OPEN, FINISHED, CANCELLED] }
        configuration: { $ref: '#/components/schemas/ProductionConfigurationRequest' }
        events: { type: array, items: { $ref: '#/components/schemas/ProductionEventResponse' } }
        finishedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ProductionGroupResponse:
      type: object
      additionalProperties: false
      required: [id, version, groupNo, buyerOrderId, members]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 0 }
        groupNo: { type: string }
        buyerOrderId: { type: string, format: uuid }
        members: { type: array, items: { $ref: '#/components/schemas/ProductionGroupMemberResponse' } }
    ProductionFinishResponse:
      type: object
      additionalProperties: false
      required: [productionOrder, stockPosition]
      properties:
        productionOrder: { $ref: '#/components/schemas/ProductionOrderResponse' }
        stockPosition: { $ref: '#/components/schemas/StockPositionResponse' }
...
    ProductionOrderPageResponse:
      type: object
      additionalProperties: false
      required: [items, page, filters, sort]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ProductionOrderResponse' } }
        page: { $ref: '#/components/schemas/PageMetadata' }
        filters: { $ref: '#/components/schemas/AppliedFilters' }
        sort: { $ref: '#/components/schemas/AppliedSort' }
    ProductionEventPageResponse:
      type: object
      additionalProperties: false
      required: [items, page, filters, sort]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ProductionEventResponse' } }
        page: { $ref: '#/components/schemas/PageMetadata' }
        filters: { $ref: '#/components/schemas/AppliedFilters' }
        sort: { $ref: '#/components/schemas/AppliedSort' }
...